third_party/WebKit/LayoutTests/fast/dom/shadow/querySelector-with-detached-node-distribution.html

<!DOCTYPE html>
<head>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body></body>
<script>
description('crbug.com/337618: querySelector needs ensure distribution.');

var host = createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('div', {'class': 'foo'},
                createDOM('content', {'id': 'content', 'select': '*'}))),
        createDOM('div', {'id': 'child'},
            createShadowRoot(
                createDOM('div', {'id': 'bar'},
                     document.createTextNode("bar")))));

var barShadowRoot = host.querySelector('#child').shadowRoot;
shouldBe('barShadowRoot.querySelectorAll(":host-context(.foo)::shadow div")[0].id', '"bar"');

var content = host.shadowRoot.querySelector('#content');
content.setAttribute('select', '.foobar');
shouldBe('barShadowRoot.querySelectorAll(":host-context(.foo)::shadow div")[0]', "undefined");
</script>
